integrate basebackup in storage model
All backend specific functions from basebackup.go were migrated to the storage
Extending the storagreinterface
