+++ backup/basebackup.go
@@ -24,19 +24,10 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"regexp"
 	"sort"
-	"strings"
 	"text/tabwriter"
-	"time"
-
-	log "github.com/Sirupsen/logrus"
 
 	"github.com/dustin/go-humanize"
-	minio "github.com/minio/minio-go"
 )
 
 // IsSane returns false if at leased one backups seams not sane
@@ -69,68 +60,6 @@ func (b *Backups) Insane() (insane Backups) {
 	return insane
 }
 
-// AddFile adds a new backup to Backups
-func (b *Backups) AddFile(path string) (err error) {
-	var newBackup Backup
-	// Make a relative path absolute
-	newBackup.Path, err = filepath.Abs(path)
-	if err != nil {
-		return err
-	}
-	newBackup.Extension = filepath.Ext(path)
-	// Get the name of the backup file without the extension
-	newBackup.Name = strings.TrimSuffix(filepath.Base(path), newBackup.Extension)
-
-	// Get size of backup
-	file, err := os.Open(path)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	fi, err := file.Stat()
-	if err != nil {
-		return err
-	}
-	newBackup.Size = fi.Size()
-
-	// Remove anything before the '@'
-	reg := regexp.MustCompile(`.*@`)
-	backupTimeRaw := reg.ReplaceAllString(newBackup.Name, "${1}")
-	newBackup.Created, err = time.Parse(BackupTimeFormat, backupTimeRaw)
-	if err != nil {
-		return err
-	}
-	// Add back reference to the list od backups
-	newBackup.Backups = b
-	b.Backup = append(b.Backup, newBackup)
-	b.Sort()
-	return nil
-}
-
-// AddObject adds a new backup to Backups
-func (b *Backups) AddObject(object minio.ObjectInfo, bucket string) (err error) {
-	var newBackup Backup
-	newBackup.Bucket = bucket
-	newBackup.Extension = filepath.Ext(object.Key)
-
-	// Get Name without suffix
-	newBackup.Name = strings.TrimSuffix(object.Key, newBackup.Extension)
-	newBackup.Size = object.Size
-
-	// Get the time from backup name
-	backupTimeRaw := extractTimeFromBackup.ReplaceAllString(newBackup.Name, "${1}")
-	newBackup.Created, err = time.Parse(BackupTimeFormat, backupTimeRaw)
-	if err != nil {
-		return err
-	}
-	// Add back reference to the list of backups
-	newBackup.Backups = b
-	b.Backup = append(b.Backup, newBackup)
-	b.Sort()
-	return nil
-}
-
 // String returns an overview of the backups as string
 func (b *Backups) String() (backups string) {
 	buf := new(bytes.Buffer)
@@ -150,42 +79,6 @@ func (b *Backups) String() (backups string) {
 	fmt.Fprintln(w, "Total backups:", b.Len(), " Not sane backups:", notSane)
 	w.Flush()
 	return buf.String()
-}
-
-// GetBackupsInDir includes all backups in given directory
-func (b *Backups) GetBackupsInDir(backupDir string) {
-	files, _ := ioutil.ReadDir(backupDir)
-	for _, f := range files {
-		err := b.AddFile(backupDir + "/" + f.Name())
-		if err != nil {
-			log.Warn(err)
-		}
-	}
-	// Sort backups
-	b.Sort()
-}
-
-// GetBackupsInBucket includes all backups in given bucket
-func (b *Backups) GetBackupsInBucket(bucket string) {
-	// Create a done channel to control 'ListObjects' go routine.
-	doneCh := make(chan struct{})
-	defer close(doneCh)
-
-	isRecursive := true
-	objectCh := b.MinioClient.ListObjects(bucket, "", isRecursive, doneCh)
-	for object := range objectCh {
-		if object.Err != nil {
-			log.Error(object.Err)
-		}
-		log.Debug(object)
-		err := b.AddObject(object, bucket)
-		if err != nil {
-			log.Error(err)
-		}
-	}
-
-	// Sort backups
-	b.Sort()
 }
 
 // Backups implements sort.Interface based on Backup.Created
@@ -210,69 +103,6 @@ func (b *Backups) SortDesc() {
 func (b *Backups) SortAsc() {
 	// Sort, the newest first
 	sort.Sort(b)
-}
-
-// SeparateBackupsByAge separates the backups by age
-// The newest "countNew" backups are put in newBackups
-// The older backups which are not already in newBackups are put in oldBackups
-func (b *Backups) SeparateBackupsByAge(countNew uint) (newBackups Backups, oldBackups Backups, err error) {
-	// Sort backups first
-	b.SortDesc()
-
-	// If there are not enough backups, return all as new
-	if (*b).Len() < int(countNew) {
-		return *b, Backups{}, errors.New("Not enough new backups")
-	}
-
-	// Give the additional vars to the new sets
-	newBackups.MinioClient = b.MinioClient
-	oldBackups.MinioClient = b.MinioClient
-	newBackups.WalDir = b.WalDir
-	oldBackups.WalDir = b.WalDir
-	newBackups.WalBucket = b.WalBucket
-	oldBackups.WalBucket = b.WalBucket
-
-	// Put the newest in newBackups
-	newBackups.Backup = (b.Backup)[:countNew]
-
-	// Put every other backup in oldBackups
-	oldBackups.Backup = (b.Backup)[countNew:]
-
-	if newBackups.IsSane() != true {
-		return newBackups, oldBackups, errors.New("Not all backups (newBackups) are sane" + newBackups.String())
-	}
-
-	if newBackups.Len() <= 0 && oldBackups.Len() > 0 {
-		panic("No new backups, only old. Not sane! ")
-	}
-	return newBackups, oldBackups, nil
-}
-
-// DeleteAll deletes all backups in the struct
-func (b *Backups) DeleteAll() (count int, err error) {
-	// Sort backups
-	b.SortDesc()
-	// We delete all backups, but start with the oldest just in case
-	for i := len(b.Backup) - 1; i >= 0; i-- {
-		backup := b.Backup[i]
-		if backup.Path != "" {
-			err = os.Remove(backup.Path)
-			if err != nil {
-				log.Warn(err)
-			} else {
-				count++
-			}
-		}
-		if backup.Bucket != "" {
-			err = b.MinioClient.RemoveObject(backup.Bucket, backup.Name+backup.Extension)
-			if err != nil {
-				log.Warn(err)
-			} else {
-				count++
-			}
-		}
-	}
-	return count, err
 }
 
 // Find finds a backup by name and returns it
