+++ storage/backends/local/local.go
@@ -21,11 +21,16 @@
 package local
 
 import (
+	"errors"
 	"io"
+	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
+	"strings"
 	"sync"
+	"time"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/xxorde/pgglaskugel/backup"
@@ -39,8 +44,17 @@ type Localbackend struct {
 // GetBackups returns backups
 func (b Localbackend) GetBackups(viper func() map[string]interface{}, subDirWal string) (backups backup.Backups) {
 	log.Debug("Get backups from folder: ", viper()["backupdir"])
-	backups.GetBackupsInDir(viper()["backupdir"].(string))
-	backups.WalDir = filepath.Join(viper()["archivedir"].(string), subDirWal)
+	backupDir := viper()["backupdir"].(string)
+	files, _ := ioutil.ReadDir(backupDir)
+	for _, f := range files {
+		err := addFileToBackups(backupDir+"/"+f.Name(), &backups)
+		if err != nil {
+			log.Warn(err)
+		}
+	}
+	// Sort backups
+	backups.Sort()
+
 	return backups
 }
 
@@ -170,4 +184,89 @@ func (b Localbackend) GetBasebackup(viper func() map[string]interface{}, backup 
 	wgDone.Wait()
 	log.Debug("getFromFile done")
 
+}
+
+// SeparateBackupsByAge separates the backups by age
+// The newest "countNew" backups are put in newBackups
+// The older backups which are not already in newBackups are put in oldBackups
+func (b Localbackend) SeparateBackupsByAge(countNew uint, backups *backup.Backups) (newBackups backup.Backups, oldBackups backup.Backups, err error) {
+	// Sort backups first
+	backups.SortDesc()
+
+	// If there are not enough backups, return all as new
+	if (*backups).Len() < int(countNew) {
+		return *backups, backup.Backups{}, errors.New("Not enough new backups")
+	}
+
+	// Put the newest in newBackups
+	newBackups.Backup = (backups.Backup)[:countNew]
+
+	// Put every other backup in oldBackups
+	oldBackups.Backup = (backups.Backup)[countNew:]
+
+	if newBackups.IsSane() != true {
+		return newBackups, oldBackups, errors.New("Not all backups (newBackups) are sane" + newBackups.String())
+	}
+
+	if newBackups.Len() <= 0 && oldBackups.Len() > 0 {
+		panic("No new backups, only old. Not sane! ")
+	}
+	return newBackups, oldBackups, nil
+}
+
+// DeleteAll deletes all backups in the struct
+func (b Localbackend) DeleteAll(backups *backup.Backups) (count int, err error) {
+	// Sort backups
+	backups.SortDesc()
+	// We delete all backups, but start with the oldest just in case
+	for i := len(backups.Backup) - 1; i >= 0; i-- {
+		backup := backups.Backup[i]
+		err = os.Remove(backup.Path)
+		if err != nil {
+			log.Warn(err)
+		} else {
+			count++
+		}
+
+	}
+	return count, err
+}
+
+// AddFile adds a new backup to Backups
+func addFileToBackups(path string, b *backup.Backups) (err error) {
+	var newBackup backup.Backup
+	// Make a relative path absolute
+	newBackup.Path, err = filepath.Abs(path)
+	if err != nil {
+		return err
+	}
+	newBackup.Extension = filepath.Ext(path)
+	// Get the name of the backup file without the extension
+	newBackup.Name = strings.TrimSuffix(filepath.Base(path), newBackup.Extension)
+
+	// Get size of backup
+	file, err := os.Open(path)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	fi, err := file.Stat()
+	if err != nil {
+		return err
+	}
+	newBackup.Size = fi.Size()
+
+	// Remove anything before the '@'
+	reg := regexp.MustCompile(`.*@`)
+	backupTimeRaw := reg.ReplaceAllString(newBackup.Name, "${1}")
+	newBackup.Created, err = time.Parse(backup.BackupTimeFormat, backupTimeRaw)
+	if err != nil {
+		return err
+	}
+	// Add back reference to the list of backups
+	newBackup.Backups = b
+	b.Backup = append(b.Backup, newBackup)
+	b.Sort()
+	return nil
 }
